backend/requirements.txt
# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Snowflake
snowflake-connector-python>=3.6.0

# Database (App DB) - uncomment if using PostgreSQL
# sqlalchemy>=2.0.25
# alembic>=1.13.1
# asyncpg>=0.29.0
# psycopg2-binary>=2.9.9

# Redis - uncomment if using Redis
# redis>=5.0.1
# hiredis>=2.3.2

# Azure - uncomment if using Azure
# azure-identity>=1.15.0
# azure-keyvault-secrets>=4.7.0
# azure-storage-blob>=12.19.0

# OAuth
httpx>=0.27.0

# JWT
python-jose[cryptography]>=3.3.0

# Monitoring
sentry-sdk[fastapi]>=1.40.0

# Utilities
python-multipart>=0.0.9
python-dotenv>=1.0.0
